+++ R/add_n.R
@@ -80,7 +80,13 @@ add_n.tbl_summary <- function(x, statistic = "{n}", col_label = "**N**", footnot
     map_dfr(
       function(.x) {
         df_stats <-
-          select(.x, any_of(c(
+          .x %>%
+          # remove overall row, if it has been added with `add_overall()`
+          purrr::when(
+            "by" %in% names(.) ~ filter(., !is.na(.data$by)),
+            TRUE ~ .
+          ) %>%
+          select(any_of(c(
             "variable", "by", "N_obs", "N_miss", "N_nonmiss", "p_miss",
             "p_nonmiss", "N_obs_unweighted", "N_miss_unweighted",
             "N_nonmiss_unweighted", "p_miss_unweighted",